// src/api-client/models/business-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Kind of business, which shapes expected volume and logistics.
 * @export
 * @enum {string}
 */

export const BusinessType = {
    Supermarket: 'SUPERMARKET',
    Hypermarket: 'HYPERMARKET',
    Restaurant: 'RESTAURANT',
    Hotel: 'HOTEL',
    Bakery: 'BAKERY',
    Catering: 'CATERING',
    Distributor: 'DISTRIBUTOR',
    Manufacturer: 'MANUFACTURER',
    Farm: 'FARM',
    CorporateCafeteria: 'CORPORATE_CAFETERIA',
    ConvenienceStore: 'CONVENIENCE_STORE',
    Other: 'OTHER'
} as const;

export type BusinessType = typeof BusinessType[keyof typeof BusinessType];