// src/api-client/models/recipient-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * What kind of receiver this is.
 * @export
 * @enum {string}
 */

export const RecipientType = {
    Ngo: 'NGO',
    FoodBank: 'FOOD_BANK',
    SoupKitchen: 'SOUP_KITCHEN',
    Shelter: 'SHELTER',
    CommunityFridge: 'COMMUNITY_FRIDGE',
    Church: 'CHURCH',
    School: 'SCHOOL'
} as const;

export type RecipientType = typeof RecipientType[keyof typeof RecipientType];